+++ inc/Smartling/ApiWrapperInterface.php
@@ -48,7 +48,7 @@ interface ApiWrapperInterface
     /**
      * @throws SmartlingApiException
      */
-    public function acquireLock(ConfigurationProfileEntity $profile, string $key, int $ttlSeconds): \DateTime;
+    public function acquireLock(ConfigurationProfileEntity $profile, string $key, float $ttlSeconds): \DateTime;
 
     public function getSourceLocale(ConfigurationProfileEntity $profile): string;
